requirements.txt
flask
flask_cors
flask_jwt_extended
boto3
pyyaml
warrant